refactor(colonel): consolidate duplicate admin schema trees
Collapse the duplicated colonel response-schema tree that lived under both
src/schemas/api/account/responses/colonel-*.ts and
src/schemas/api/internal/responses/colonel-*.ts into a single canonical set
under internal/responses. The account-side copies were verbatim duplicates
that had drifted out of sync; the internal tree is the one the admin SPA
consumes via responseSchemas.

- Delete the 14 account/responses/colonel-*.ts duplicates and drop the
  colonel re-export from account/responses/index.ts.
- Move the canonical schema definitions into internal/responses/colonel.ts,
  sourcing systemSettings{,Details}Schema from contracts/config instead of
  the deleted account copy.
- Repoint every admin view/store import at the internal tree.
- Fix useSystemSettings / systemSettingsStore to import systemSettingsSchema
  and SystemSettingsDetails from contracts/config.

Type-check, lint, and the full schema + admin vitest suites (88 files,
1665 tests) pass.
