+++ src/apps/admin/stores/useAdminOrganizations.ts
@@ -4,7 +4,7 @@ import {
   usePaginatedFetch,
   type PageMeta,
 } from '@/apps/admin/composables/usePaginatedFetch';
-import type { ColonelOrganization } from '@/schemas/api/account/responses/colonel';
+import type { ColonelOrganization } from '@/schemas/api/internal/responses/colonel';
 import { colonelOrganizationsResponseSchema } from '@/schemas/api/internal/responses/colonel';
 import { defineStore } from 'pinia';
 import { ref } from 'vue';
