+++ src/schemas/api/account/responses/index.ts
@@ -1,5 +1,4 @@
 // src/schemas/api/account/responses/index.ts
 
 export * from './account';
-export * from './colonel';
 export * from './permissions';
